Add TCP force-close test client implementation
- Implemented a TCP client that connects to a server and sends messages.
- Added functions for parsing IP addresses, sending and receiving messages, and handling errors.
- Included regex validation for coefficient messages.
- Implemented logic to test server disconnection behavior with a small receive buffer.
- Added command-line argument parsing for server IP, port, IP type, and expected coefficient response.

+++ proj2/src/messages.h
@@ -183,7 +183,7 @@ static inline types::rational_t deserialize_offset(const std::string& str) {
         if (!std::regex_match(str, regex)) {
             throw std::invalid_argument("Invalid offset format");
         }
-        return static_cast<types::rational_t>(std::stof(str));
+        return static_cast<types::rational_t>(std::stod(str));
     } catch (const std::exception&) {
         throw std::invalid_argument("Invalid offset value");
     }
